Hello world works!!!
In lib.blarb in this commit, "Hello, world!" will be printed!

Believe it or not, this was a BIG job in BlarbLang! Now that string and
heap operations are available, it should be _relatively_ easy to read
input from files, the console, write to files and the console, networks,
MAGIC! BLACK MAGIC! IT'S ALL BLACK MAGIC!!!