[#3] One step closer to eliminating malloc
The heap memory allocation should be much more stable now. It should be
fine to use the heap as much as you like, as long as you don't load new
blarb files while using the heap.

I am toying with the idea of shoving all loaded symbols into a static
buffer.. but that will be implemented another day.